chore: refactor FeeEdit and FeeItems
