+++ designer/client/components/Fee/FeeItems.jsx
@@ -237,3 +237,5 @@ export class FeeItems extends React.Component {
     );
   }
 }
+FeeItems.value = undefined;
+
